// src/storyblok/types/component-types-sb.d.ts
export interface BlockConfigStoryblok {
  vertical_whitespace_top: "large" | "medium" | "small" | "none";
  vertical_whitespace_bottom: "large" | "medium" | "small" | "none";
  component: "block_config";
  _uid: string;
}

export interface ButtonStoryblok {
  label?: string;
  variant?: "primary" | "outline";
  component: "button";
  _uid: string;
}
// ...
export interface RichtextStoryblok {
  type: string;
  content?: RichtextStoryblok[];
  marks?: RichtextStoryblok[];
  attrs?: unknown;
  text?: string;
}
// ...
export type MultiassetStoryblok = {
  alt: string | null;
  copyright?: string | null;
  fieldtype: "asset";
  id: number;
  filename: string | null;
  name: string;
  title: string | null;
  focus: string | null;
  meta_data?: {};
  source?: string | null;
  is_external_url?: boolean;
  is_private?: boolean;
  src?: string;
  updated_at?: string;
  width?: number | null;
  height?: number | null;
  aspect_ratio?: number | null;
  public_id?: string | null;
  content_type?: string;
}[];

export interface FeatureBlockStoryblok {
  config: BlockConfigStoryblok[];
  eyebrow_title?: string;
  title?: string;
  body?: RichtextStoryblok;
  ctas?: ButtonStoryblok[];
  images?: MultiassetStoryblok;
  theme?: number | string;
  view_type?: "feature_style_1" | "feature_style_2" | "feature_style_3";
  component: "feature_block";
  _uid: string;
}